backend/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.4.0
Pillow==10.4.0
stripe==7.8.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
whitenoise==6.5.0
dj-database-url==2.1.0